refactor: move ImageTile to components/common
ImageTile is a generic background-picker tile, now used by both the
ayat-hadith designer and the custom theme controls. Move it out of the
designer feature into components/common and import via the barrel.
