+++ src/components/common/image-tile.tsx
