+++ src/components/common/index.ts
@@ -5,3 +5,4 @@ export * from './action-buttons';
 export * from './data-table';
 export * from './orientation-badge';
 export * from './auth-legal-links';
+export * from './image-tile';
